Missing channel attribute to bitstamp.Event data struct
Problem:

When subscribing to multiple channels there was no way to know from
which channel the published/received data was coming from.

We may want to subscribe to multiple channels, for example
"live_trades_btcusd", "live_trades_ethbtc", etc.

We need a way to know from which channel the message is coming from.